4-clear_bit.c

+++ 0x14-bit_manipulation/4-clear_bit.c
@@ -1,5 +1,4 @@
 #include "main.h"
-#include <stddef.h>
 
 /**
  * clear_bit - sets the value of a bit to 0 at a given index.
@@ -16,7 +15,7 @@ int clear_bit(unsigned long int *n, unsigned int index)
 	mask = 1;
 	mask = mask << index;
 
-	if (index > sizeof(unsigned int) * 8 || n == NULL)
+	if (index > sizeof(unsigned long int) * 8 || n == NULL)
 		return (-1);
 	if (((*n >> index) & 1) == 1)
 		*n = mask ^ *n;
